// shared/types.ts
export interface Product {
  id: string;
  name: string;
  description: string;
  price: number;
  costPrice: number; // Preço de custo do produto
  stock: number;
  lowStockThreshold: number;
  categoryId: string; // Alterado de 'category' para 'categoryId'
  barcode: string;
  imageUrl: string;
  // Dados fiscais
  ncm: string;
  cest: string;
  cfop: string;
  origin: string;
}

// Representa um cliente
export interface Customer {
  id:string;
  name: string;
  cpfCnpj: string;
  email: string;
  phone: string;
  address?: string; // Endereço para entregas
}

// Representa um item em uma venda
export interface SaleItem {
  productId: string;
  productName: string;
  quantity: number;
  returnableQuantity: number; // Quantidade que ainda pode ser devolvida
  unitPrice: number;
  totalPrice: number;
  imageUrl?: string;
}